move matcher to separate file
